requirements.txt
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.25.1
psycopg2-binary==2.9.7
asyncpg==0.28.0
SQLAlchemy==2.0.23
sqlalchemy-utils==0.41.1
alembic==1.12.0
redis==5.0.0
asyncio_redis==0.16.0
anyio==3.7.1
alembic==1.12.0
fastapi-users==12.1.2
email-validator==2.0.0
python-jose==3.3.0
auth0-python==4.4.2